refactor: separate `testSolutions` function to `solution.ts`
